Add break-glass cron buttons to /admin (#110)
The owner needed CRON_SECRET to manually kick the cron during the 2026-05-02
incident, and didn't have it saved — wasted ~10 min rotating before recovery
could start. /admin already runs under a stronger auth check (ADMIN_EMAIL +
session) than the bearer token, so add server actions there that invoke the
same code paths as /api/cron and /api/cron/schedule. Recovery time drops from
~10 min to ~30 sec.

- Extract the cron logic into lib/cron-jobs.js (runMainCron, runScheduler) so
  both the existing route handlers and the new server actions share one
  implementation.
- New server actions in app/admin/actions.js re-check ADMIN_EMAIL server-side
  (defense-in-depth — page-level notFound() is not the security boundary).
- New client component app/admin/break-glass.js renders the buttons and shows
  the response inline via useActionState.
- Document the new break-glass path in CLAUDE.md and INCIDENT.md as the
  primary recovery path; bearer-token curl is now the fallback.
